+++ scripts/externalImage.js
@@ -1,0 +1,97 @@
+function appendQueryParams(url, params) {
+    const { searchParams } = url;
+    params.forEach((value, key) => {
+      searchParams.set(key, value);
+    });
+    url.search = searchParams.toString();
+    return url.toString();
+  }
+  export function createOptimizedPicture(src, alt = '', eager = false, breakpoints = [{ media: '(min-width: 600px)', width: '2000' }, { width: '750' }]) {
+    const isAbsoluteUrl = /^https?:\/\//i.test(src);
+  
+    // Fallback to createOptimizedPicture if src is not an absolute URL
+    if (!isAbsoluteUrl) return libCreateOptimizedPicture(src, alt, eager, breakpoints);
+  
+    const url = new URL(src);
+    const picture = document.createElement('picture');
+    const { pathname } = url;
+    const ext = pathname.substring(pathname.lastIndexOf('.') + 1);
+  
+    // webp
+    breakpoints.forEach((br) => {
+      const source = document.createElement('source');
+      if (br.media) source.setAttribute('media', br.media);
+      source.setAttribute('type', 'image/webp');
+      const searchParams = new URLSearchParams({ width: br.width, format: 'webply' });
+      source.setAttribute('srcset', appendQueryParams(url, searchParams));
+      picture.appendChild(source);
+    });
+  
+    // fallback
+    breakpoints.forEach((br, i) => {
+      const searchParams = new URLSearchParams({ width: br.width, format: ext });
+  
+      if (i < breakpoints.length - 1) {
+        const source = document.createElement('source');
+        if (br.media) source.setAttribute('media', br.media);
+        source.setAttribute('srcset', appendQueryParams(url, searchParams));
+        picture.appendChild(source);
+      } else {
+        const img = document.createElement('img');
+        img.setAttribute('loading', eager ? 'eager' : 'lazy');
+        img.setAttribute('alt', alt);
+        picture.appendChild(img);
+        img.setAttribute('src', appendQueryParams(url, searchParams));
+      }
+    });
+  
+    return picture;
+  }
+  
+  function getUrlExtension(url) {
+    return url.split(/[#?]/)[0].split('.').pop().trim();
+  }
+  function isExternalImage(element, externalImageMarker) {
+    // if the element is not an anchor, it's not an external image
+    if (element.tagName !== 'A') return false;
+  
+    // if the element is an anchor with the external image marker as text content,
+    // it's an external image
+    if (element.textContent.trim() === externalImageMarker) {
+      return true;
+    }
+  
+    // if the element is an anchor with the href as text content and the href has
+    // an image extension, it's an external image
+    if (element.textContent.trim() === element.getAttribute('href')) {
+      const ext = getUrlExtension(element.getAttribute('href'));
+      return ext && ['jpg', 'jpeg', 'png', 'gif', 'webp'].includes(ext.toLowerCase());
+    }
+  }
+  export function decorateExternalImages(ele, deliveryMarker) {
+    const extImages = ele.querySelectorAll('a');
+    extImages.forEach((extImage) => {
+      if (isExternalImage(extImage, deliveryMarker)) {
+        const extImageSrc = extImage.getAttribute('href');
+        const extPicture = createOptimizedPicture(extImageSrc);
+        console.log("Inside the function");
+        /* copy query params from link to img */
+        const extImageUrl = new URL(extImageSrc);
+        const { searchParams } = extImageUrl;
+        extPicture.querySelectorAll('source, img').forEach((child) => {
+          if (child.tagName === 'SOURCE') {
+            const srcset = child.getAttribute('srcset');
+            if (srcset) {
+              child.setAttribute('srcset', appendQueryParams(new URL(srcset, extImageSrc), searchParams));
+            }
+          } else if (child.tagName === 'IMG') {
+            const src = child.getAttribute('src');
+            if (src) {
+              child.setAttribute('src', appendQueryParams(new URL(src, extImageSrc), searchParams));
+            }
+          }
+        });
+        extImage.parentNode.replaceChild(extPicture, extImage);
+      }
+    });
+  }
